Move `BlockReason` to Talpid package
