+++ mullvad-jni/src/lib.rs
@@ -35,14 +35,6 @@ const CLASSES_TO_LOAD: &[&str] = &[
     "java/net/InetSocketAddress",
     "java/util/ArrayList",
     "net/mullvad/mullvadvpn/model/AccountData",
-    "net/mullvad/mullvadvpn/model/BlockReason$AuthFailed",
-    "net/mullvad/mullvadvpn/model/BlockReason$Ipv6Unavailable",
-    "net/mullvad/mullvadvpn/model/BlockReason$SetFirewallPolicyError",
-    "net/mullvad/mullvadvpn/model/BlockReason$SetDnsError",
-    "net/mullvad/mullvadvpn/model/BlockReason$StartTunnelError",
-    "net/mullvad/mullvadvpn/model/BlockReason$ParameterGeneration",
-    "net/mullvad/mullvadvpn/model/BlockReason$IsOffline",
-    "net/mullvad/mullvadvpn/model/BlockReason$TapAdapterProblem",
     "net/mullvad/mullvadvpn/model/AppVersionInfo",
     "net/mullvad/mullvadvpn/model/Constraint$Any",
     "net/mullvad/mullvadvpn/model/Constraint$Only",
@@ -84,6 +76,14 @@ const CLASSES_TO_LOAD: &[&str] = &[
     "net/mullvad/talpid/tunnel/ActionAfterDisconnect$Block",
     "net/mullvad/talpid/tunnel/ActionAfterDisconnect$Nothing",
     "net/mullvad/talpid/tunnel/ActionAfterDisconnect$Reconnect",
+    "net/mullvad/talpid/tunnel/BlockReason$AuthFailed",
+    "net/mullvad/talpid/tunnel/BlockReason$Ipv6Unavailable",
+    "net/mullvad/talpid/tunnel/BlockReason$SetFirewallPolicyError",
+    "net/mullvad/talpid/tunnel/BlockReason$SetDnsError",
+    "net/mullvad/talpid/tunnel/BlockReason$StartTunnelError",
+    "net/mullvad/talpid/tunnel/BlockReason$ParameterGeneration",
+    "net/mullvad/talpid/tunnel/BlockReason$IsOffline",
+    "net/mullvad/talpid/tunnel/BlockReason$TapAdapterProblem",
     "net/mullvad/talpid/tunnel/ParameterGenerationError$NoMatchingRelay",
     "net/mullvad/talpid/tunnel/ParameterGenerationError$NoMatchingBridgeRelay",
     "net/mullvad/talpid/tunnel/ParameterGenerationError$NoWireguardKey",
